FODO SC: 3D case with 1/10 sync motion WORKS!
... sorted out longitudinal synchrotron motion (missing rvv, rpp and psigma
variable updates in SixTrackLib) and the newly created mesh was too small,
now it covers all 4 quadrants (and not only until the centre of the beam!).
This latter one was introduced only when going to Ji's new parameters when
setting up the grid by myself. (I forgot a factor 2)

emittance growth is identical to Ji's IMPACT simulations!