SIS100: comparing with RF between PySixTrack, SixTrackLib and MAD-X
The ipynb documents the systematic difference between SixTrackLib and MAD-X
when switching on RF and with a certain non-vanishing longitudinal coordinate.

This current version avoids any longitudinal momentum setter problems on the PySixTrack / SixTrackLib side
by using the from_mad_track method to copy the correct values from MAD-X for comparison.

<~>

1. I find no systematic offset between the zeta drift in PySixTrack and MAD-X any more.
The difference oscillates around 0 and decreases during one turn.

2. I find 1e-8 transverse error comparing long-term tracking over 20k turns between
PySixTrack and MAD-X. ==> good!

3. I find a considerable error comparing the same between SixTrackLib and MAD-X because of
the cavity issue in SixTrackLib.

In the lower part of the ipynb, the difference of the cavity kick is illustrated during one turn
when comparing SixTrackLib and PySixTrack.

==> this cavity kick needs to be fixed in SixTrackLib, might have to do with the setters of
the longitudinal momentum coordinates.